services/python/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
playwright==1.47.0
beautifulsoup4==4.12.3
lxml==5.3.0
supabase==2.7.4
openai==1.47.0
python-dotenv==1.0.1
pydantic==2.9.2
apscheduler==3.10.4
tenacity==9.0.0
pypdf==5.0.1
langchain-text-splitters==0.3.0
tqdm==4.66.5
